Moving patch-airsense.py into new directory
